Move monad methods directly to Result
